Reduce issue loading time
This commit implements TVar to store the loaded issues
so they don't need to queried ad-hoc. Correspondingly,
this commit simplifies the elm logic so we don't have
to cache the issues client side anymore. However, if
a user wants to explicitly refresh the issues without
loading the page, they can use the `refresh` endpoint.